refactored genesis state values into a separate trait and objects
