+++ src/main/resources/application.conf
@@ -2,7 +2,7 @@ bifrost {
 
   restApi {
     # Local network address to bind to
-    bindAddress = "127.0.0.1:9085"
+    bindAddress = "0.0.0.0:9085"
 
     # Hash of the API key string
     apiKeyHash = ""
