+++ src/main/scala/co/topl/http/api/ApiRouteWithView.scala
@@ -3,7 +3,7 @@ package co.topl.http.api
 import akka.actor.ActorRef
 import akka.pattern.ask
 import co.topl.nodeView.CurrentView
-import co.topl.nodeView.GenericNodeViewHolder.ReceivableMessages.GetDataFromCurrentView
+import co.topl.nodeView.NodeViewHolder.ReceivableMessages.GetDataFromCurrentView
 import co.topl.nodeView.history.History
 import co.topl.nodeView.mempool.MemPool
 import co.topl.nodeView.state.State
